+++ HISTORY.rst
@@ -7,3 +7,13 @@ History
 ++++++++++++++++++
 
 * First release as a module.
+
+1.1 (2014-05-21)
+++++++++++++++++++
+
+* Proper initialization of curses colors
+
+1.2 (2014-05-21)
+++++++++++++++++
+
+* Added support for 256 color terminals mimicking the original web game